Move random carousel to landing folder
