+++ site/gatsby-site/src/components/landing/RandomIncidentsCarousel.js
